@property directive compilation pipeline changed to allow control base type generation in VS Extension
